Move model/validator.py to common/validator.py
