+++ featurebyte/models/scd_data.py
@@ -7,9 +7,9 @@ from typing import Any, ClassVar, List, Type
 
 from pydantic import root_validator
 
+from featurebyte.common.validator import construct_data_model_root_validator
 from featurebyte.enum import DBVarType
 from featurebyte.models.feature_store import DataModel
-from featurebyte.models.validator import construct_data_model_root_validator
 from featurebyte.query_graph.model.common_table import BaseTableData
 from featurebyte.query_graph.model.table import SCDTableData
