+++ featurebyte/schema/tabular_data.py
@@ -8,10 +8,10 @@ from typing import List, Optional
 from bson.objectid import ObjectId
 from pydantic import Field, StrictStr, validator
 
+from featurebyte.common.validator import columns_info_validator
 from featurebyte.models.base import FeatureByteBaseModel, PydanticObjectId
 from featurebyte.models.feature_store import DataStatus
 from featurebyte.models.tabular_data import TabularDataModel
-from featurebyte.models.validator import columns_info_validator
 from featurebyte.query_graph.model.column_info import ColumnInfo
 from featurebyte.query_graph.model.common_table import TabularSource
 from featurebyte.schema.common.base import BaseDocumentServiceUpdateSchema, PaginationMixin
